refactor(components): refactor autocomplete
