+++ packages/components/autocomplete/src/autocomplete.ts
@@ -1,0 +1,78 @@
+import { NOOP } from '@vue/shared'
+import { buildProps, definePropType } from '@element-plus/utils/props'
+import { UPDATE_MODEL_EVENT } from '@element-plus/utils/constants'
+import { isString, isObject } from '@element-plus/utils/util'
+import type { ExtractPropTypes } from 'vue'
+
+import type { Placement } from '@element-plus/components/popper'
+
+export const autocompleteProps = buildProps({
+  valueKey: {
+    type: String,
+    default: 'value',
+  },
+  modelValue: {
+    type: [String, Number],
+    default: '',
+  },
+  debounce: {
+    type: Number,
+    default: 300,
+  },
+  placement: {
+    type: definePropType<Placement>(String),
+    values: [
+      'top',
+      'top-start',
+      'top-end',
+      'bottom',
+      'bottom-start',
+      'bottom-end',
+    ],
+    default: 'bottom-start',
+  },
+  fetchSuggestions: {
+    type: definePropType<
+      (queryString: string, cb: (data: any[]) => void) => void
+    >(Function),
+    default: NOOP,
+  },
+  popperClass: {
+    type: String,
+    default: '',
+  },
+  triggerOnFocus: {
+    type: Boolean,
+    default: true,
+  },
+  selectWhenUnmatched: {
+    type: Boolean,
+    default: false,
+  },
+  hideLoading: {
+    type: Boolean,
+    default: false,
+  },
+  popperAppendToBody: {
+    type: Boolean,
+    default: true,
+  },
+  highlightFirstItem: {
+    type: Boolean,
+    default: false,
+  },
+} as const)
+
+export type AutocompleteProps = ExtractPropTypes<typeof autocompleteProps>
+
+export const autocompleteEmits = {
+  [UPDATE_MODEL_EVENT]: (value: string) => isString(value),
+  input: (value: string) => isString(value),
+  change: (value: string) => isString(value),
+  focus: (evt: FocusEvent) => evt instanceof FocusEvent,
+  blur: (evt: FocusEvent) => evt instanceof FocusEvent,
+  clear: () => true,
+  select: (item: object) => isObject(item),
+}
+
+export type AutocompleteEmits = typeof autocompleteEmits
